Translate website by URL
It was a prototype for a project which was using Bing translator API. I
was trying to translate a entire weisite into another language ie.
English to Chinese. However, there is a limitation for each transction.
Besides, it can only transfer 1million character each month. So,
basiclly, it is not a very good way to achieve my goal. Anyway, i think
that there is still something for me to investigate in future.